+++ app/src/main/java/io/dnajd/presentation/project_table/components/ProjectTableContent.kt
@@ -20,7 +20,7 @@ fun ProjectTableContent(
     onCreateTableTaskMenuClicked: (Long) -> Unit,
     onTaskClicked: (Long) -> Unit,
     onSwapTablePositionsClicked: (Long, Long) -> Unit,
-    onSwitchDropdownMenuClicked: (Int?) -> Unit,
+    onSwitchDropdownMenuClicked: (Long?) -> Unit,
 ) {
     Row(
         modifier = Modifier
